[feeds.conf] Fix #351
This commit fixes issue #351. The file feeds.conf included in qMp, which
holds the OpenWrt feeds other than the defaults needed to create the
binary firmware images, contains an invalid period character (".") for OpenWrt's
14.07 Barrier Breaker old packages feed's identifier. This is not accepted
by OpenWrt's firmware building, resulting in an error when the feeds are
updated and installed before actually compiling. Changing the period to an
underscore fixes it.