[qmp-system] Fix #377 (1): backport QinQ branch into testing
This commit fixes issue #377, regarding the ways VLANs are specified in
/etc/config/network. The previous method (with @wlan0-like stuff) was never
meant to work, although it just happened to work. Recent OpenWrt changes
have deprecated that *unsupported* way to declare VLANs. Backporting the
work in branch QinQ and adding a couple of changes (e.g. linklayer details
for BMX6) fixes the problem.